refactor(agent-tools): SEC-007 split grep and shell by responsibility, not by raising a limit
Adding containment pushed `grep-tool.ts` to 323 lines and `shell-tool.ts` to 309
against the 300-line anti-monolith limit, and pushed `globFileTool` and
`file-read`'s `executeWithConfig` past the 50-line function limit. The baseline
is not regenerated and no limit is raised — the cause is fixed, as SEC-006 did
when it hit the same ratchet.

Each split is by responsibility, so the two halves change for different reasons:

- `grep-search.ts` — HOW the search is performed (enumeration + per-file
  matching); `grep-tool.ts` keeps the tool SURFACE (schema, description,
  factory, result envelope).
- `shell-tool-description.ts` — the model-facing description CONTRACT (NEUT-002
  neutral default text a consumer overrides at the composition root);
  `shell-tool.ts` keeps process execution.
- `resolveSearchRoot` in `path-guard.ts` — the root resolution Glob and Grep had
  each spelled out, which also removes the duplicate rationale comment.
- `containmentDenial` in each dag file node, with the rationale moved to the
  class doc where a reader looks for it.

No behaviour change: agent-tools 239, pack-coding 11, agent-framework 1302,
dag-node-file-read 8, dag-node-file-write 9. file-size scan passes with the
baseline untouched.
