+++ packages/agent-tools/src/builtins/grep-tool.ts
@@ -13,13 +13,13 @@
  * at least as strictly as `Read`, which it could otherwise stand in for.
  */
 
-import { readFile, readdir, stat } from 'node:fs/promises';
-import { join, resolve } from 'node:path';
+import { readFile, stat } from 'node:fs/promises';
 
 import pLimit from 'p-limit';
 import { z } from 'zod';
 
-import { checkPathWithinCwd, isWithinCwd } from './path-guard.js';
+import { collectFiles, searchFile } from './grep-search.js';
+import { resolveSearchRoot } from './path-guard.js';
 import { createZodFunctionTool } from '../implementations/function-tool';
 
 import type { IContainedBuiltinToolOptions } from './tool-options.js';
@@ -65,125 +65,6 @@ type TGrepArgs = z.infer<typeof GrepSchema>;
 /** Cap on concurrent file reads during the content scan (CLI-042). */
 const READ_CONCURRENCY_LIMIT = 50;
 
-/** Convert a simple glob to a RegExp for file name filtering. */
-function globToRegex(glob: string): RegExp {
-  const escaped = glob
-    .replace(/[.+^${}()|[\]\\]/g, '\\$&')
-    .replace(/\*\*/g, '.+')
-    .replace(/\*/g, '[^/]*');
-  return new RegExp(`^${escaped}$`);
-}
-
-/** Check if a file name matches an optional glob filter. */
-function matchesGlob(filename: string, glob: string | undefined): boolean {
-  if (glob === undefined) return true;
-  return globToRegex(glob).test(filename);
-}
-
-/**
- * Gather all files under a directory recursively, excluding node_modules/.git.
- *
- * `containmentRoot` (SEC-007) drops any entry whose CANONICAL path escapes the root, before it is
- * descended into or read. `stat` follows symlinks, so without this a link inside the root pointing
- * out of it made the whole target tree readable — including, for a symlinked FILE, its contents.
- */
-async function collectFiles(
-  dirPath: string,
-  glob: string | undefined,
-  containmentRoot: string | undefined,
-): Promise<string[]> {
-  const results: string[] = [];
-
-  async function walk(current: string): Promise<void> {
-    let entryNames: string[];
-    try {
-      entryNames = await readdir(current);
-    } catch {
-      return;
-    }
-
-    for (const name of entryNames) {
-      if (name === 'node_modules' || name === '.git') continue;
-
-      const fullPath = join(current, name);
-      if (!isWithinCwd(fullPath, containmentRoot)) continue;
-      let fileStat: Awaited<ReturnType<typeof stat>>;
-      try {
-        fileStat = await stat(fullPath);
-      } catch {
-        continue;
-      }
-
-      if (fileStat.isDirectory()) {
-        await walk(fullPath);
-      } else if (fileStat.isFile()) {
-        if (matchesGlob(name, glob)) {
-          results.push(fullPath);
-        }
-      }
-    }
-  }
-
-  await walk(dirPath);
-  return results;
-}
-
-/** Search a single file for lines matching the regex. */
-function searchFile(
-  content: string,
-  filePath: string,
-  regex: RegExp,
-  contextLines: number,
-  outputMode: 'files_with_matches' | 'content' | 'count',
-): string[] {
-  const lines = content.split('\n');
-  const matchingIndices: number[] = [];
-
-  for (let i = 0; i < lines.length; i++) {
-    if (regex.test(lines[i])) {
-      matchingIndices.push(i);
-    }
-  }
-
-  if (matchingIndices.length === 0) return [];
-
-  if (outputMode === 'files_with_matches') {
-    return [filePath];
-  }
-
-  if (outputMode === 'count') {
-    return [`${filePath}:${matchingIndices.length}`];
-  }
-
-  // content mode — include context lines
-  const includedIndices = new Set<number>();
-  for (const idx of matchingIndices) {
-    for (
-      let c = Math.max(0, idx - contextLines);
-      c <= Math.min(lines.length - 1, idx + contextLines);
-      c++
-    ) {
-      includedIndices.add(c);
-    }
-  }
-
-  const outputLines: string[] = [];
-  const sortedIndices = Array.from(includedIndices).sort((a, b) => a - b);
-
-  let prevIdx: number | undefined;
-  for (const idx of sortedIndices) {
-    if (prevIdx !== undefined && idx > prevIdx + 1) {
-      outputLines.push('--');
-    }
-    const lineNum = idx + 1;
-    const marker = matchingIndices.includes(idx) ? ':' : '-';
-    outputLines.push(`${filePath}:${lineNum}${marker}${lines[idx]}`);
-    prevIdx = idx;
-  }
-
-  return outputLines;
-}
-
 async function grepFileTool(
   args: TGrepArgs,
   options: IContainedBuiltinToolOptions,
@@ -197,12 +78,7 @@ async function grepFileTool(
     headLimit,
   } = args;
   const containmentRoot = options.cwd;
-  // As in Glob: a relative `path` anchors to the containment root, so the guard root and the search
-  // root are never two different directories.
-  const searchBase = containmentRoot ?? process.cwd();
-  const targetPath = searchPath ? resolve(searchBase, searchPath) : searchBase;
-
-  const rootError = checkPathWithinCwd(targetPath, containmentRoot);
+  const { root: targetPath, error: rootError } = resolveSearchRoot(searchPath, containmentRoot);
   if (rootError) return rootError;
 
   let regex: RegExp;
